fix: Add download result purging after completion to clear memory and clutter
clear_results is now available as a config option. If set to true, all
download results (completed/error/removed) will be purged from aria2c's
memory

Adding download result purging required access to the aria2c xmlrpc client
object in the download_all method. Instead of creating the client object
every time we add a URI, it is now created once in the __init__ method.

The same is now done for the put.io client object, and it is no longer being
passed between methods. There is also more cleanup in the __init__, which
now always assigns attributes, even if they're values weren't passed as
arguments. This allows API access to the PutioAria2cDownloader object to work
similarly to the CLI with defaults

This commit also fixes a small bug where even if you passed a root_watch_dir,
it wasn't ever honored and always treated 0 as the root watch directory

fix #1